feat: abstract gtm and htmlparser into separate files, add wip directory migrator
